Tweak top/side bar layout and organize content
- Move websocket and message handler setup back to main component
- Move styles belonging to main/header to those respective files
- Remove duplicates between side bar and profile dropdown
- Remove chevron icon and ensure hamburger button is always visible
  to collapse or expand side bar
- Slightly expand text of about page
