+++ static/js/components/ProfileDropdown.jsx
@@ -39,33 +39,9 @@ const ProfileDropdown = () => {
 
           <div className={styles.rule} />
 
-          <Link to="/candidates" role="link">
-            <div className={styles.entry} onClick={collapseDropdown} role="menuitem" tabIndex="-1">
-              Scan Candidates
-            </div>
-          </Link>
-
-          <div className={styles.rule} />
-
-          <Link to="/sources" role="link">
-            <div className={styles.entry} onClick={collapseDropdown} role="menuitem" tabIndex="-1">
-              Sources
-            </div>
-          </Link>
-
-          <div className={styles.rule} />
-
           <Link to="/groups" role="link">
             <div role="menuitem" tabIndex="-1" className={styles.entry} onClick={collapseDropdown}>
               Groups
-            </div>
-          </Link>
-
-          <div className={styles.rule} />
-
-          <Link to="/about" role="link">
-            <div role="menuitem" tabIndex="-1" className={styles.entry} onClick={collapseDropdown}>
-              About
             </div>
           </Link>
