+++ static/js/components/SidebarAndHeader.jsx
@@ -2,27 +2,88 @@ import React from "react";
 import PropTypes from "prop-types";
 import { useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
-import clsx from "clsx";
+
 import Drawer from "@material-ui/core/Drawer";
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import List from "@material-ui/core/List";
-import Divider from "@material-ui/core/Divider";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
 import { blue } from "@material-ui/core/colors";
 import HomeIcon from "@material-ui/icons/Home";
 import SearchIcon from "@material-ui/icons/Search";
-import AccountBoxIcon from "@material-ui/icons/AccountBox";
 import InfoIcon from "@material-ui/icons/Info";
 import MenuIcon from "@material-ui/icons/Menu";
 import IconButton from "@material-ui/core/IconButton";
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
+import { makeStyles } from "@material-ui/core/styles";
 
-import useStyles from "./muiStyles";
 import HeaderContent from "./HeaderContent";
 import * as Actions from "../ducks/sidebar";
+
+
+const drawerWidth = 190;
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: 'flex',
+  },
+  appBar: {
+    position: "fixed",
+    zIndex: 100,
+    transition: theme.transitions.create(['margin', 'width'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen,
+    }),
+    height: "6em",
+    background: "#38B0DE",
+  },
+  appBarShift: {
+    width: `calc(100% - ${drawerWidth}px)`,
+    marginLeft: drawerWidth,
+    transition: theme.transitions.create(['margin', 'width'], {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  },
+  menuButton: {
+    marginRight: theme.spacing(2),
+    paddingTop: "1.3em",
+  },
+  hide: {
+    display: 'none',
+  },
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+  },
+  drawerPaper: {
+    zIndex: 99,
+    width: drawerWidth,
+    paddingLeft: "0.4em",
+    background: "#33345C",
+    fontSize: "1.2em",
+  },
+  toolbar: {
+    display: "flex",
+    height: "4em",
+    padding: "1em 0em",
+    alignItems: "center",
+  },
+  drawerHeader: {
+    display: 'flex',
+    alignItems: 'center',
+    padding: theme.spacing(0, 1),
+    // necessary for content to be below app bar
+    ...theme.mixins.toolbar,
+    paddingTop: "6em",
+    justifyContent: 'flex-end',
+  },
+  link: {
+    color: "#B8D2FF",
+    textDecoration: "none",
+  }
+}));
 
 
 const SidebarAndHeader = ({ open, root }) => {
@@ -36,10 +97,7 @@ const SidebarAndHeader = ({ open, root }) => {
   return (
     <>
       <AppBar
-        position="fixed"
-        className={clsx(classes.appBar, {
-          [classes.appBarShift]: open,
-        })}
+        className={classes.appBar}
       >
         <Toolbar>
           <IconButton
@@ -47,7 +105,7 @@ const SidebarAndHeader = ({ open, root }) => {
             aria-label="open drawer"
             onClick={handleToggleSidebarOpen}
             edge="start"
-            className={clsx(classes.menuButton, open && classes.hide)}
+            className={classes.menuButton}
           >
             <MenuIcon />
           </IconButton>
@@ -63,12 +121,7 @@ const SidebarAndHeader = ({ open, root }) => {
           paper: classes.drawerPaper,
         }}
       >
-        <div className={classes.drawerHeader}>
-          <IconButton onClick={handleToggleSidebarOpen}>
-            <ChevronLeftIcon />
-          </IconButton>
-        </div>
-        <Divider />
+        <div className={classes.drawerHeader} />
         <List>
           <Link to="/" className={classes.link}>
             <ListItem button name="sidebarDashboardButton">
@@ -84,14 +137,6 @@ const SidebarAndHeader = ({ open, root }) => {
                 <SearchIcon style={{ color: blue[200] }} />
               </ListItemIcon>
               <ListItemText primary="Candidates" />
-            </ListItem>
-          </Link>
-          <Link to="/profile" className={classes.link}>
-            <ListItem button name="sidebarProfileButton">
-              <ListItemIcon>
-                <AccountBoxIcon style={{ color: blue[200] }} />
-              </ListItemIcon>
-              <ListItemText primary="Profile" />
             </ListItem>
           </Link>
           <Link to="/about" className={classes.link}>
