--- static/js/components/muiStyles.js
@@ -1,85 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-
-const drawerWidth = 190;
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-  },
-  appBar: {
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    height: "6em",
-    background: "#38B0DE",
-  },
-  appBarShift: {
-    width: `calc(100% - ${drawerWidth}px)`,
-    marginLeft: drawerWidth,
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-    paddingTop: "1.3em",
-  },
-  hide: {
-    display: 'none',
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-  },
-  drawerPaper: {
-    width: drawerWidth,
-    paddingLeft: "0.4em",
-    background: "#33345C",
-    fontSize: "1.2em",
-  },
-  toolbar: {
-    display: "flex",
-    height: "4em",
-    padding: "1em 0em",
-    alignItems: "center",
-  },
-  drawerHeader: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    // necessary for content to be below app bar
-    ...theme.mixins.toolbar,
-    justifyContent: 'flex-end',
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(2),
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginTop: "6em", // top bar height
-    marginLeft: drawerWidth,
-    "& a": {
-      textDecoration: "none",
-      color: "gray",
-      fontWeight: "bold",
-    },
-  },
-  contentShift: {
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    marginLeft: 0,
-  },
-  link: {
-    color: "#B8D2FF",
-    textDecoration: "none",
-  }
-}));
-
-export default useStyles;
